Finished bot logging/db insertion

+++ db_control.py
@@ -13,10 +13,11 @@ def run_commit(sql,conn,curs):
 def creat_tQuote(conn, curs):
     sql = """ CREATE TABLE IF NOT EXISTS tQuote (
                 id integer PRIMARY KEY,
-                logger_name text,
-                quoter_name text,
+                logger_name text NOT NULL,
+                quotee_name text,
                 quote text NOT NULL,
-                date text
+                context text,
+                date DATETIME DEFAULT CURRENT_TIMESTAMP
             ); """
     run_commit(sql, conn, curs)
 
@@ -37,10 +38,23 @@ def backup_log(quote):
         f.write(str(quote) + '\n')
 
 def insert_quote(conn, curs, quote):
-    sql = ''' INSERT INTO tQuote(logger_name,quoter_name,quote,date)
+    sql = ''' INSERT INTO tQuote(logger_name,quotee_name,quote,context)
               VALUES(?,?,?,?)'''
     try:
         curs.execute(sql, quote)
+        # curs.execute('INSERT INTO tQuote (LastModifiedTime) VALUES(CURRENT_TIMESTAMP)')
+        conn.commit()
+    except Error as e:
+        print(e)
+    backup_log(quote)
+    return curs.lastrowid
+
+def insert_quote_small(conn, curs, quote):
+    sql = ''' INSERT INTO tQuote(logger_name,quote)
+              VALUES(?,?)'''
+    try:
+        curs.execute(sql, quote)
+        # curs.execute('INSERT INTO tQuote (LastModifiedTime) VALUES(CURRENT_TIMESTAMP)')
         conn.commit()
     except Error as e:
         print(e)
@@ -56,29 +70,30 @@ if __name__ == "__main__":
     elif (sys.argv[1] == '-help') or (sys.argv[1] == '-h'):
         print("""Command List:
         -h, -help      :  Review command options
-        -rb, -rebuild  :  DELTES and rebuilds the QuotesDB database
+        -rb, -rebuild  :  DELTES and rebuilds the QuoteDB database
         -v, -view      :  Prints all entries in the database
         """)
     else:
-        db_name = 'Quotes.db'
+        db_name = 'Quote.db'
         if (sys.argv[1] == '-rb') or (sys.argv[1] == '-rebuild'):
             inp = input("Are you sure you would like to delete the existing database? (Y/N): ")
             if inp.lower() == 'n':
                 print('Phew, close call')
                 sys.exit()
             print("Rebuilding database")
-            os.remove(db_name)
+            if os.path.exists(db_name):
+                os.remove(db_name)
             conn = connect_db(db_name)
             if (not conn):
-                print('Failed to build database, exiting')
+                print('Failed to rebuild database, exiting')
             else:
                 curs = conn.cursor()
                 build_tables(conn, curs)
-                print("Successfully rebuilt Quotes.db database")
+                print("Successfully rebuilt Quote.db database")
         elif (sys.argv[1] == '-v') or (sys.argv[1] == '-view'):
             conn = connect_db(db_name)
             if (not conn):
-                print('Failed to build database, exiting')
+                print('Failed to connect to database, exiting')
             else:
                 curs = conn.cursor()
                 # print(pd.read_sql("SELECT * FROM sqlite_master;",conn))
@@ -86,10 +101,10 @@ if __name__ == "__main__":
         elif(sys.argv[1] == '-t') or (sys.argv[1] == '-test'):
             conn = connect_db(db_name)
             if (not conn):
-                print('Failed to build database, exiting')
+                print('Failed to connect to database, exiting')
             else:
                 curs = conn.cursor()
-                quote = ('Wils', 'Lee', 'Eatin ass and eatin ass', '2020-11-22')
+                quote = ('Wils', 'Lee', 'Eatin ass and eatin ass', 'Triggs referring to a canibal orgy')
                 print(insert_quote(conn, curs, quote))
         conn.close()
